requirements.txt
streamlit
google-generativeai
Pillow
scikit-image
pandas
numpy